Interaction Point check toggle
added a bool to toggle between checking for interaction points and ignoring that requirement. Now we have the option for the Agent to only require sim objects to be Visible to interact with them, rather than requiring visibility AND hand access to the object's interaction points.

This might be useful for cases where a sink or cabinet is at a weird angle causing the hand to not have line of sight to interaction points since the Hand and Camera are at different positions, causing inconsistent raycasts