Simplified Visible/Interactable Check, Also Fridges
Physics Fridge scripts added so you can see the fridge as "visible" even with the fridge doors are open. Getting rid of redundant "Interactable" checks and now only checking if an object is Visible to be actionable. Added lots of debug spheres to help check visibility points as well as bounding box locations.