Restructure of Cabinet/Drawer Heirarchy to allow for raycasting invisible
Restructure of how the CanOpen.cs script works, switching over to the new CanOpen_Object script that is more universal. Also found that this new structure requires all non trigger colliders to have an additional collider with a rigidbody that is NOT linked to the "my colliders" of the sim object. This rbCol is only there to ensure proper collision when hitting other sim objects. It seems that when animating the root rigidbody made up of compound colliders, there are no issues, but animating just a group of colliders and NOT the root rigidbody causes the rigidbody physics to not propogate down to the child/grandchild colliders correctly. What a pain.