Transparent Sim Objects
So transparent sim objects are weird and we probably want to revisit this implementation at some point cuase it's quite something. All box colliders must be set to tag: SimObjPhysics and layer:SimObjInvisible, but each collider needs a copy of itself parented to it that is set to Tag:untagged, and Layer:TransparentObjects to allow non transparent objects to be visible through transparent ones. However, there is an issue where this "translucent" raycast can only be sent one object deep, so if multiple translucent objects are in front of each other you won't be able to see through all of them