All Kitchens set dressed with Hide and Seek Objects
Also fixed a visibility check typo so now we can use the raycastAll version (yay!). tweaked the actionFinished check in Debug Controller to delay for a little bit to make sure all actionFinished calls have time to finish (this should be removed once we find all the issues cause it's pretty hacky).

Changed the "CanOpen" test button to the equals sign because hitting 1 was being used for other things.