Updated CheckIfAgentCanRotateHand()
caught an edge case in the hand item rotation check. Redid the logic to also allow for use of compound collider creating tool that I found. Leaving original implementation commented out for now in case something breaks and we need to switch back. That should't be the case though.

Also added some local collider making tools on my computer to ignore file so it's not pushed onto the repo.