Additional spawn parameter for spawning in Agent hand or not
If we are spawning an object in the agent's hand, we don't need to check for collision with the Agent. If we are spawning it in the environment directly, we need to check for both hitting structures AND hitting the agent so objects aren't spawned inside the agent.

Basically objects can only spawn inside the agent if they are going to immediately be placed in the agent's hand.