Adjusted Rotate checker heirarchy on Agent...
to prevent look up/look down from having the incorrectly rotated rotate box checks when crouching. Because crouching moved the position of the camera but did not move the rotation checking boxes relative to the camera, the angles were incorrect. Now they are paranted to the camera so everything remains correctly relative.

Also went through all the old scenes and cleaned up the "Targets" and "Controls" empty transforms that were there for some reason. Not sure what they were used for but they aren't needed for physics.