done with PlaceObjectAtPoint() action
action takes a SimObj that exists in the scene by UniqueId, and tries to teleport it to the given point (action.position).

Best when used with GetReachablePositions or some other function that can return world space coordinates.

This action does not attempt to rotate the object that is being placed, so it will only succeed if the object's current rotation can fit in the position specified.