adding toggle between Bot and Tall mode
adding new parameter to Initialize action: agentMode, which can be set to "Tall" or "Bot" to switch between the normal Tall agent and the smaller Locobot mode in any scene.

The Default is on Initialize, the Tall mode will activate.

Also removing references to SSAO since that is being finnicky at the moment.