adjusted action.fieldOfView defaults and initialization
removed references to action.fov since it is redundant, now all methods requiring field of view inputs use action.fieldOfView

action.fieldOfView now defaults to 0, and on initialize the agentMode value determines if it defaults to 90 or 42.5 based on the mode (tall or bot). On Initialize, values can overwrite the 90/42.5 defaults if passed in. The ChangeFOV action can also overwrite current for without running Initialize again.