fix to early call of ActionFinished with ToggleOn/Off
actionFinished() was being returned immediately when an toggleable object had finished animating.

When an object is toggled, after animation tweening is done, setIsOn() is called to complete all state changes in order to turn on the object. Because the coroutine delaying actionFinished only delayed until all tweening was done, it would return early by a frame even though not all properties of the toggelable object had finished changing.