fix toward ScreenToWorldTarget()
When casting against the default layer (0<<1) in ScreenToWorldTarget, there are cases where a collider on an object like Cabinets and Drawers are on the Default layer, causing the raycast to be blocked from certain angles, resulting in a false "Object not found!" error even though the sim object is within the unobstructed path of the ray.

Adding a check to check the collider hit by the ray and all parents of that object to see if it is a sim object in order to account for this.