Fix to RemoveFromScene objectId assignment
instead of re-assigning objectId every time an object is removed from the scene, only the key for that object is removed from the objectIdToSimObjPhysics dict.

This prevents issues with objectId being re-assigned incorrectly due to floating point errors.

Additionally, moved DisableAllObjectsOfType to base controller to be in line with the EnableObject() and DisableObject() actions. Added notes to these actions for when a more comprehensive update of the objectId assignment logic is revamped.