Adjustments to the custom objects flow
The idea is to allow the TA1 researchers to use the custom registry file, set as a file reference, to refer to objects anywhere under a Resources folder.

Set the file reference to a dummy registry file I created, and adjusted code paths so as to load that reference. Then adjusted the way that objects are created off of the registry to allow the function to use the full path for addressables, yet ignore the addressables path prefix if it's not an addressable.

I should move this to separate functions, to be clearer on what it does. The code to read Resources assets in the AddressablesUtil script definitely feels wrong.