tweak to visibility check
adding edge case where if a raycast to a visibility point reaches the point and hits no colliders, the object should be visible.

This checks for cases like if the agent camera is casting a ray from inside a trigger collider of an object, which normally fails to report visibility since raycasts from inside colliders can't collide with those colliders.