Maintain contained objects when placed on stove
Without this, a pan with a potato in it placed on the stove will no longer
logically contain the potato, even though it's clearly sitting there. The code,
including the comment about why we use the stationary drop, is copied exactly
from the `PlaceObject()`, where I observed it working fine for placing the pan
with the potato in it onto the counter.